+++ CHANGES.rst
@@ -17,6 +17,7 @@
 * utils.make_name now takes caller-specified counter numbers.
 * sqlalchemy.BaseNameMixin and BaseScopedNameMixin.make_name now take a reserved names list.
 * New: utils.nullint for returning int or None.
+* New: utils.nullstr for returning unicode or None.
 
 0.3.13
 ------